adding more components
